minor bugfixes and improvements

+++ include/villas/log.h
@@ -89,7 +89,8 @@ struct log {
 	int level;
 
 	/** Debug facilities used by the debug() macro. */
-	int facilities;
+	long facilities;
+	
 };
 
 /** Initialize log object */
